spinner componentify
