+++ frontend/src/pages/home/components/UpcomingMeeting.tsx
@@ -11,6 +11,7 @@ import ConfirmModal from '../../../utility_components/ConfirmModal/ConfirmModal'
 import NewMeeting from './NewMeeting';
 import { useUserContext } from '../../../contexts/UserContext';
 import { useMeetingContext } from '../../../contexts/MeetingContext';
+import LoadingDots from '../../../utility_components/Loading/LoadingDots';
 
 type UpcommingMeetingProps = {
     meeting: MeetingDTO;
@@ -61,18 +62,7 @@ export default function UpcomingMeeting({ meeting }: UpcommingMeetingProps) {
     }
 
     if (isLoading) {
-        return (
-            <div className="fixed z-50 top-0 left-0 w-full h-full overflow-hidden bg-gray-800 opacity-50 flex items-center justify-center">
-                <div className="bg-white border py-2 px-5 rounded-lg flex flex-col items-center">
-                    <div className="loader-dots grid grid-cols-3 gap-2">
-                        <div className="bg-[#7d6ca3] rounded-full w-2 h-2"></div>
-                        <div className="bg-[#7d6ca3] rounded-full w-2 h-2"></div>
-                        <div className="bg-[#7d6ca3] rounded-full w-2 h-2"></div>
-                    </div>
-                    <div className="text-gray-500 text-xs font-light mt-2 text-center">Please wait...</div>
-                </div>
-            </div>
-        );
+        return <LoadingDots />;
     }
 
     return (
